+++ Models/Interfaces/IEllipticalWing.cs
@@ -8,11 +8,11 @@ public interface IEllipticalWing
 	/// <summary>
 	/// Gets or sets the sweep coefficient. Defines the curvature of the elliptical wing.
 	/// </summary>
-	public float SweepCoefficient { get; set; }
+	public double SweepCoefficient { get; set; }
 	
 	/// <summary>
 	/// Gets or sets the ratio relative to the wingspan
 	/// to exclude at the tip of the wing in millimeters.
 	/// </summary>
-	public float TipExclusionRatio { get; set; }
+	public double TipExclusionRatio { get; set; }
 }
