+++ Models/Interfaces/ITaperedWing.cs
@@ -14,5 +14,5 @@ public interface ITaperedWing
 	/// Gets or sets the taper ratio of the wing, which is
 	/// the ratio of the tip chord length to the root chord length.
 	/// </summary>
-	public float TaperRatio { get; set; }
+	public double TaperRatio { get; set; }
 }
